cleanup and added test
