+++ network/transport/v2/handlers.go
@@ -270,58 +270,6 @@ func (p *protocol) handleGossip(peer transport.Peer, envelope *Envelope) error {
 	return nil
 }
 
-func (p *protocol) handleTransactionList(peer transport.Peer, envelope *Envelope) error {
-	subEnvelope := envelope.Message.(*Envelope_TransactionList)
-	msg := envelope.GetTransactionList()
-	cid := conversationID(msg.ConversationID)
-	data := handlerData{}
-
-	log.Logger().Tracef("handling handleTransactionList from peer (peer=%s, conversationID=%s, message=%d/%d)", peer.ID.String(), cid.String(), msg.MessageNumber, msg.TotalMessages)
-
-	// check if response matches earlier request
-	if _, err := p.cMan.check(subEnvelope, data); err != nil {
-		return err
-	}
-
-	txs, err := subEnvelope.parseTransactions(data)
-	if err != nil {
-		return err
-	}
-
-	p.mLock.Lock()
-	defer p.mLock.Unlock()
-
-	refsToBeRemoved := map[hash.SHA256Hash]bool{}
-
-	ctx := context.Background()
-	maxLC := uint32(0)
-	for i, tx := range txs {
-		// TODO does this always trigger fetching missing payloads? (through observer on DAG) Prolly not for v2
-		if len(tx.PAL()) == 0 && len(msg.Transactions[i].Payload) == 0 {
-			return fmt.Errorf("peer did not provide payload for transaction (tx=%s)", tx.Ref())
-		}
-		if err = p.state.Add(ctx, tx, msg.Transactions[i].Payload); err != nil {
-			if errors.Is(err, dag.ErrPreviousTransactionMissing) {
-				p.cMan.done(cid)
-				log.Logger().Warnf("ignoring remainder of TransactionList due to missing prevs (conversation=%s, Tx with missing prevs=%s)", cid, tx.Ref())
-				xor, clock := p.state.XOR(ctx, math.MaxUint32)
-				return p.sender.sendState(peer.ID, xor, clock)
-			}
-			return fmt.Errorf("unable to add received transaction to DAG (tx=%s): %w", tx.Ref(), err)
-		}
-		if tx.Clock() > maxLC {
-			maxLC = tx.Clock()
-		}
-		refsToBeRemoved[tx.Ref()] = true
-	}
-
-	if msg.MessageNumber >= msg.TotalMessages {
-		p.cMan.done(cid)
-	}
-
-	return nil
-}
-
 func (p *protocol) handleTransactionListQuery(peer transport.Peer, envelope *Envelope) error {
 	msg := envelope.GetTransactionListQuery()
 	requestedRefs := make([]hash.SHA256Hash, len(msg.Refs))
